Use getFloat instead getInteger for "chance" at ExtratableItemsData.
- Thanks to Henrique S (https://www.l2jbrasil.com/topic/138662-l2jorion/?do=findComment&comment=844193&_rid=1392)

+++ L2jOrion_SourceCode/src/l2jorion/game/datatables/xml/ExtractableItemsData.java
@@ -62,7 +62,7 @@ public class ExtractableItemsData implements IXmlReader
 							
 							production = set.getInteger("id");
 							amount = set.getInteger("quantity");
-							chance = set.getInteger("chance");
+							chance = set.getFloat("chance");
 							extractables.add(new L2ExtractableProductItem(production, amount, chance));
 							totalChance = 0;
 							for (L2ExtractableProductItem extractable : extractables)
